fix msgpack_options

+++ msgpack.h
@@ -11,7 +11,7 @@ namespace HPHP {
 
 static Array HHVM_FUNCTION(msgpack_unpack, const String& data);
 static String HHVM_FUNCTION(msgpack_pack, const Array& data);
-static void HHVM_FUNCTION(msgpack_set_options, const int options);
+static void HHVM_FUNCTION(msgpack_set_options, const int64_t options);
 
 
 class MsgpackExtension : public Extension {
